Remove tabs from GNexusParts
Although the tabs are pretty and shows off some of the nice features
of Androud ICS/JB they are against the design guidelines.  Having
a tab for a single item is inefficient and it doesn't match the
rest of the settings.

This moves all the tabs to PreferenceCategories and have all
settings on one page.

Changeset 2+3+4+5: Fix whitespace and tab fail.

Change-Id: I36fe8b1611ccc345ceab534b01f4dc9a97fef70c

Conflicts:
	GNexusParts/res/xml/display_preferences.xml
	GNexusParts/src/com/cyanogenmod/settings/device/DeviceSettings.java
	GNexusParts/src/com/cyanogenmod/settings/device/DisplayFragmentActivity.java

Conflicts:

	GNexusParts/src/com/cyanogenmod/settings/device/DeviceSettings.java